屈世超 Redid+接口加密

+++ Model/DealRecord.cs
@@ -13,7 +13,7 @@ namespace Model
         [Key]
         public int Did { get; set; }
         public DateTime Dtime { get; set; }
-        public double Dmoney { get; set; }
+        public decimal Dmoney { get; set; }
         public string Oid { get; set; }
         public int Cid { get; set; }
         public int Uid { get; set; }
